hotfix: unlink chain account
